Remove fallback PIDDIR_USER from Debian init script
The removed id -u check caused a bug where the init script would fail
if the PIDDIR didn't already exist. PIDDIR_USER was never being set to
PIDDIR_USER_FALLBACK if it was empty because 'id -u' doesn't return >0

Then the install command would fail as it was trying to interpret '-g'
as a user name.